move zarrita data generation function to a separate file
remove zarrita code from zarr
