+++ generate_data/generate_zarr.py
@@ -1,5 +1,4 @@
 import zarr
-import zarrita
 import numcodecs
 from skimage.data import astronaut
 
@@ -44,25 +43,6 @@ def generate_n5_format(compressors=['gzip', None]):
                          compressor=compressor_impl)
 
 
-def generate_zr3_format(compressors=['gzip', 'blosc', 'zlib', None]):
-    im = astronaut()
-
-    h = zarrita.create_hierarchy('data/zarr.zr3')
-    for compressor in compressors:
-        copts = COMPRESSION_OPTIONS.get(compressor, {})
-        if compressor is None:
-            name = "raw"
-        elif compressor == "blosc":
-            name = "%s/%s" % (compressor, copts.get("cname"))
-        else:
-            name = compressor
-        compressor_impl = STR_TO_COMPRESSOR[compressor](**copts) if compressor is not None else None
-        a = h.create_array('/' + name, shape=im.shape, chunk_shape=CHUNKS,
-                           dtype=im.dtype, compressor=compressor_impl)
-        a[...] = im
-
-
 if __name__ == '__main__':
     generate_zarr_format()
     generate_n5_format()
-    generate_zr3_format()
